chore: F144 -- remove unwired pre-architecture Android bg-scan code (4 files + workmanager dep); ScanFrequency extracted; design verdict written, F161 registered (#333)
