+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Format: `- **type**: Description (Issue #N)` where type is feat|fix|chore|docs
 ## [Unreleased]
 
 ### 2026-08-16 (Sprint 60)
+- **chore**: F144 -- the unwired pre-architecture Android background-scan code is REMOVED: `BackgroundScanManager`, `BackgroundScanService`, `background_scan_worker.dart`, and `BackgroundScanNotificationService` had zero call sites (superseded by the Windows Task Scheduler architecture, ADR-0039/0040), and the `workmanager` dependency went with them. The live `ScanFrequency` enum (used by the Windows scheduler path and Settings) extracted to `scan_frequency.dart` with its tests consolidated. Design verdict written: a full Windows-mirroring Android scheduler is feasible (WorkManager per-account unique periodic work maps 1:1) and is registered as F161 (~3-5h) together with the POST_NOTIFICATIONS runtime request -- deliberately deferred with it, since no Android notification call site exists yet to gate the permission prompt on (premise corrected during evaluation). (Issue #333)
 - **feat**: F143 -- the Review No Rule Items screen gains a touch selection model: long-press a row to select it (entering selection mode), and on touch platforms a plain tap then toggles rows in and out of the selection (touch's Ctrl+click) with the existing selection bar as the contextual action surface. Desktop semantics are strictly unchanged (replace-single click, Ctrl/Shift, right-click menu) and pinned by an explicit regression-guard test. The AppBar Review No Rule Items icon loses its Windows gate (F142's own deferral note pointed here), so Android gets the deep-link entry point on non-default screens; the gate-pin tests now assert the icon on every platform including CI's ubuntu runner. Platform is read via Theme.of(context).platform for a per-tree test seam. All changes mutation-verified. (Issue #332)
 - **test**: F159 -- the version-consistency gate (Dart test + CLI script) now also asserts `msix_config.msix_version` equals the pubspec version + `.0` (it drifted silently at 0.6.0.0 for ~10 releases with nothing watching). Both paths mutation-verified red/green. Recorded scope decision: dependency-floor claims in pubspec comments vs constraints remain deliberately unwatched (no robust cheap check without a new comment convention); that class stays a review-time concern. (Issue #331)
 - **test**: F158 -- CI gains an Android Build Verification job (`flutter build apk --debug`, ubuntu + JDK 17), closing the F150 regression class (Android build broken for months, discovered only manually). The gitignored `google-services.json` is satisfied in CI by a committed STUB (`android/ci/google-services.ci.json`: real package name, deliberately fake everything else -- public repo, and CI never runs the app; same reasoning as F127's empty dart-defines). Stub verified locally against the exact gradle task, and mutation-verified: a wrong-package stub fails with the F150 'No matching client' signature. (Issue #330)
