+++ mobile-app/pubspec.yaml
@@ -45,7 +45,9 @@ dependencies:
   sqflite_common_ffi: ^2.3.0  # SQLite FFI for desktop platforms
 
   # Phase 3.5 Sprint 7 - Background Scanning
-  workmanager: ^0.9.0  # Background task scheduling (Android/iOS) (F108, Sprint 44: 0.5->0.9; we use NetworkType.connected + a simple inputData map, unaffected by the enum renames / native-Map change)
+  # workmanager removed F144 (Sprint 60): its only consumers were the unwired
+  # pre-architecture BackgroundScanManager/worker (zero call sites, deleted).
+  # Re-add when the Windows-mirroring Android scheduler lands (follow-up item).
   flutter_local_notifications: ^17.2.1  # Local notifications (F150, Sprint 59: 16->17; v16 does not compile against Android SDK 34 -- upstream bigLargeIcon ambiguity. NOTE: 17.0.0 only SURFACES the ambiguity (it bumps compileSdk to 34); the actual fix is upstream issue #2329, first released in 17.2.1 -- hence the floor here (Claude cowork review, 2026-08-15). We use only initialize/show/AndroidNotificationDetails; the v17 breaking changes are in scheduling APIs we do not call.)
   connectivity_plus: ^5.0.0  # Network connectivity detection
   battery_plus: ^5.0.0  # Battery level detection
