+++ mobile-app/test/unit/services/windows_task_scheduler_service_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:my_email_spam_filter/core/services/windows_task_scheduler_service.dart';
-import 'package:my_email_spam_filter/core/services/background_scan_manager.dart';
+import 'package:my_email_spam_filter/core/services/scan_frequency.dart';
 
 void main() {
   group('WindowsTaskSchedulerService', () {
